Crate graph_chain_ethereum

Re-exports

pub use crate::chain::Chain;

Modules

adapter

capabilities

chain

data_source

ethereum_adapter

network

network_indexer

runtime

transport

trigger

Structs

DataSource
Runtime representation of a data source.

EthereumAdapter

EthereumContractCall

EthereumNetworks

EventLoopHandle
Event Loop Handle. NOTE: Event loop is stopped when handle is dropped!

MockEthereumAdapter
Common trait for components that watch and manage access to Ethereum.

NodeCapabilities

ProviderEthRpcMetrics

RuntimeAdapter

SubgraphEthRpcMetrics

TriggerFilter

Enums

EthereumContractCallError

MappingTrigger

Transport
Abstraction over the different web3 transports.

Traits

EthereumAdapterTrait
Common trait for components that watch and manage access to Ethereum.

Type Definitions

DataSourceTemplate