Struct graph_chain_ethereum::adapter::MockEthereumAdapter

pub struct MockEthereumAdapter {
    EthereumAdapter_expectations: MockEthereumAdapter_EthereumAdapter,
}

Common trait for components that watch and manage access to Ethereum.

Implementations may be implemented against an in-process Ethereum node or a remote node over RPC.

Fields
EthereumAdapter_expectations: MockEthereumAdapter_EthereumAdapter
Implementations

impl MockEthereumAdapter

pub fn checkpoint(&mut self)

Validate that all current expectations for all methods have been satisfied, and discard them.

pub fn new() -> Self

Create a new mock object with no expectations.

This method will not be generated if the real struct already has a new method. However, it will be generated if the struct implements a trait with a new method. The trait’s new method can still be called like <MockX as TraitY>::new

impl MockEthereumAdapter

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_url_hostname(&mut self) -> &mut Expectation

Create an Expectation for mocking the url_hostname method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_provider(&mut self) -> &mut Expectation

Create an Expectation for mocking the provider method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_net_identifiers(&mut self) -> &mut Expectation

Create an Expectation for mocking the net_identifiers method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_latest_block(&mut self) -> &mut Expectation

Create an Expectation for mocking the latest_block method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_latest_block_header(&mut self) -> &mut Expectation

Create an Expectation for mocking the latest_block_header method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_load_block(&mut self) -> &mut Expectation

Create an Expectation for mocking the load_block method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_load_blocks(&mut self) -> &mut Expectation

Create an Expectation for mocking the load_blocks method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_block_by_hash(&mut self) -> &mut Expectation

Create an Expectation for mocking the block_by_hash method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_block_by_number(&mut self) -> &mut Expectation

Create an Expectation for mocking the block_by_number method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_load_full_block(&mut self) -> &mut Expectation

Create an Expectation for mocking the load_full_block method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_block_pointer_from_number(&mut self) -> &mut Expectation

Create an Expectation for mocking the block_pointer_from_number method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_block_hash_by_block_number(&mut self) -> &mut Expectation

Create an Expectation for mocking the block_hash_by_block_number method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_uncles(&mut self) -> &mut Expectation

Create an Expectation for mocking the uncles method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_contract_call(&mut self) -> &mut Expectation

Create an Expectation for mocking the contract_call method