Module graph_chain_ethereum::chain

Structs

Chain

DummyDataSourceTemplate

IngestorAdapter

MAX_BLOCK_RANGE_SIZE
Maximum number of blocks to request in each chunk.

TARGET_TRIGGERS_PER_BLOCK_RANGE
Ideal number of triggers in a range. The range size will adapt to try to meet this.

TriggersAdapter

Enums

BlockFinality
This is used in EthereumAdapter::triggers_in_block, called when re-processing a block for newly created data sources. This allows the re-processing to be reorg safe without having to always fetch the full block data.