Function graph_chain_ethereum::ethereum_adapter::blocks_with_triggers

pub(crate) async fn blocks_with_triggers(
    adapter: Arc<EthereumAdapter>, 
    logger: Logger, 
    chain_store: Arc<dyn ChainStore>, 
    subgraph_metrics: Arc<SubgraphEthRpcMetrics>, 
    stopwatch_metrics: StopwatchMetrics, 
    from: BlockNumber, 
    to: BlockNumber, 
    filter: &TriggerFilter, 
    unified_api_version: UnifiedMappingApiVersion
) -> Result<Vec<BlockWithTriggers<Chain>>, Error>

Returns blocks with triggers, corresponding to the specified range and filters. If a block contains no triggers, there may be no corresponding item in the stream. However the to block will always be present, even if triggers are empty.

Careful: don’t use this function without considering race conditions. Chain reorgs could happen at any time, and could affect the answer received. Generally, it is only safe to use this function with blocks that have received enough confirmations to guarantee no further reorgs, and where the Ethereum node is aware of those confirmations. If the Ethereum node is far behind in processing blocks, even old blocks can be subject to reorgs. It is recommended that to be far behind the block number of latest block the Ethereum node is aware of.