Enum graph_chain_ethereum::network_indexer::network_indexer::StateMachine

enum StateMachine {
    Start,
    EnsureSubgraph {
        ensure_subgraph: Box<dyn Future<Item = (), Error = Error> + Send>,
    },
    LoadLocalHead {
        local_head: Box<dyn Future<Item = Option<BlockPtr>, Error = Error> + Send>,
    },
    PollChainHead {
        local_head: Option<BlockPtr>,
        prev_chain_head: Option<BlockPtr>,
        chain_head: Box<dyn Future<Item = LightEthereumBlock, Error = Error> + Send>,
    },
    ProcessBlocks {
        local_head: Option<BlockPtr>,
        chain_head: LightEthereumBlock,
        next_blocks: Box<dyn Stream<Item = BlockWithOmmers, Error = Error> + Send>,
    },
    VetBlock {
        local_head: Option<BlockPtr>,
        chain_head: LightEthereumBlock,
        next_blocks: Box<dyn Stream<Item = BlockWithOmmers, Error = Error> + Send>,
        block: BlockWithOmmers,
    },
    RevertLocalHead {
        chain_head: LightEthereumBlock,
        local_head: Option<BlockPtr>,
        new_local_head: Box<dyn Future<Item = BlockPtr, Error = Error> + Send>,
    },
    AddBlock {
        chain_head: LightEthereumBlock,
        next_blocks: Box<dyn Stream<Item = BlockWithOmmers, Error = Error> + Send>,
        new_local_head: Box<dyn Future<Item = BlockPtr, Error = Error> + Send>,
    },
    Ready(()),
    Failed(Error),
}

State machine that handles block fetching and block reorganizations.

Variants

Start

The indexer start in an empty state and immediately moves on to ensuring that the network subgraph exists.

EnsureSubgraph

This state ensures that the network subgraph that stores the indexed data exists, and creates it if necessary.

Fields of EnsureSubgraph

ensure_subgraph: Box<dyn Future<Item = (), Error = Error> + Send>

LoadLocalHead

This state waits until the local head block has been loaded from the store. It then moves on to polling the chain head block.

Fields of LoadLocalHead

local_head: Box<dyn Future<Item = Option<BlockPtr>, Error = Error> + Send>

PollChainHead

This state waits until the chain head block has been polled successfully.

Based on the (local head, chain head) pair, the indexer then moves on to fetching and processing a range of blocks starting at local head + 1 up, leading up to the chain head. This is done in chunks of e.g. 100 blocks at a time for two reasons:

To limit the amount of blocks we keep in memory.

To be able to re-evaluate the chain head and check for reorgs frequently.

Fields of PollChainHead

local_head: Option<BlockPtr>prev_chain_head: Option<BlockPtr>chain_head: Box<dyn Future<Item = LightEthereumBlock, Error = Error> + Send>

ProcessBlocks

This state takes the next block from the stream. If the stream is exhausted, it transitions back to polling the chain head block and deciding on the next chunk of blocks to fetch. If there is still a block to read from the stream, it’s passed on to vetting for validation and reorg checking.

Fields of ProcessBlocks

local_head: Option<BlockPtr>chain_head: LightEthereumBlocknext_blocks: Box<dyn Stream<Item = BlockWithOmmers, Error = Error> + Send>

VetBlock

This state vets incoming blocks with regards to two aspects:

Does the block have a number and hash? This is a requirement for indexing to continue. If not, the indexer re-evaluates the chain head and starts over.

Is the block the successor of the local head block? If yes, move on to indexing this block. If not, we have a reorg.

Notes on the reorg handling:

By checking parent/child succession, we ensure that there are no gaps in the indexed data (class mathematical induction). So if the local head is x and a block f comes in that is not a successor/child, it must be on a different version/fork of the chain.

E.g.:

a---b---c---x
    \
     +--d---e---f

In that case we need to do the following:

Find the common ancestor of x and f, which is the block after which the two versions diverged (in the above example: b).

Collect old blocks betweeen the common ancestor and (including) the local head that need to be reverted (in the above example: c, x).

Fetch new blocks between the common ancestor and (including) f that are to be inserted instead of the old blocks in order to make the incoming block (f) the local head (in the above example: d, e, f).

We don’t actually do all of the above explicitly. What we do instead is that when we encounter a block that requires a reorg, we revert the local head, moving back by one block in the indexed data. We then poll the chain head again, fetch up to 100 blocks, vet the next block again, and see if we have reverted back to the common ancestor yet. If we have, we process the next block. If we haven’t, we repeat reverting the local head. Ultimately, this will take us back to the common ancestor, and at that point we can move forward again.

Fields of VetBlock

local_head: Option<BlockPtr>chain_head: LightEthereumBlocknext_blocks: Box<dyn Stream<Item = BlockWithOmmers, Error = Error> + Send>block: BlockWithOmmers

RevertLocalHead

This state reverts the local head, moving the local indexed data back by one block.

After reverting, the local head is updated to the previous local head.

If reverting fails, the local head remains unchanged. If reverting succeeds but sending out the revert event for the block fails, the local head is still moved back to its parent. At this point the block has been reverted in the store, so it’s too late. Note: This means that if an event does not arrive at one of the consumers of the indexer events, these consumers will have to reconcile what they know with what is in the store as they go.

After reverting (or failing to revert), the indexer polls the chain head again to decide what blocks to fetch and process next.

Fields of RevertLocalHead

chain_head: LightEthereumBlocklocal_head: Option<BlockPtr>new_local_head: Box<dyn Future<Item = BlockPtr, Error = Error> + Send>

AddBlock

This state waits until a block has been written and an event for it has been sent out. After that, the indexer continues processing the next block. If anything goes wrong at this point, it’s back to re-evaluating the chain head and fetching (potentially) different blocks for indexing.

Fields of AddBlock

chain_head: LightEthereumBlocknext_blocks: Box<dyn Stream<Item = BlockWithOmmers, Error = Error> + Send>new_local_head: Box<dyn Future<Item = BlockPtr, Error = Error> + Send>

Ready(())

This is unused, the indexing never ends.

Failed(Error)

State for fatal errors that cause the indexing to terminate. This should almost never happen. If it does, it should cause the entire node to crash and restart.

Implementations

impl StateMachine

fn start(context: Context) -> StateMachineFuture

Start executing the StateMachineFuture state machine. This constructing its Future representation in its initial start state and returns it.

fn start_in<STATE: Into<StateMachineStates>>(
state: STATE,
context: Context
) -> StateMachineFuture