Trait graph_chain_ethereum::network_indexer::network_indexer::PollStateMachine

trait PollStateMachine: StateMachineFuture {
    fn poll_start<'smf_poll_state, 'smf_poll_context>(
        _: &'smf_poll_state mut RentToOwn<'smf_poll_state, Start>, 
        _: &'smf_poll_context mut RentToOwn<'smf_poll_context, Context>
    ) -> Poll<AfterStart, Error>;

    fn poll_ensure_subgraph<'smf_poll_state, 'smf_poll_context>(
        _: &'smf_poll_state mut RentToOwn<'smf_poll_state, EnsureSubgraph>, 
        _: &'smf_poll_context mut RentToOwn<'smf_poll_context, Context>
    ) -> Poll<AfterEnsureSubgraph, Error>;

    fn poll_load_local_head<'smf_poll_state, 'smf_poll_context>(
        _: &'smf_poll_state mut RentToOwn<'smf_poll_state, LoadLocalHead>, 
        _: &'smf_poll_context mut RentToOwn<'smf_poll_context, Context>
    ) -> Poll<AfterLoadLocalHead, Error>;

    fn poll_poll_chain_head<'smf_poll_state, 'smf_poll_context>(
        _: &'smf_poll_state mut RentToOwn<'smf_poll_state, PollChainHead>, 
        _: &'smf_poll_context mut RentToOwn<'smf_poll_context, Context>
    ) -> Poll<AfterPollChainHead, Error>;

    fn poll_process_blocks<'smf_poll_state, 'smf_poll_context>(
        _: &'smf_poll_state mut RentToOwn<'smf_poll_state, ProcessBlocks>, 
        _: &'smf_poll_context mut RentToOwn<'smf_poll_context, Context>
    ) -> Poll<AfterProcessBlocks, Error>;

    fn poll_vet_block<'smf_poll_state, 'smf_poll_context>(
        _: &'smf_poll_state mut RentToOwn<'smf_poll_state, VetBlock>, 
        _: &'smf_poll_context mut RentToOwn<'smf_poll_context, Context>
    ) -> Poll<AfterVetBlock, Error>;

    fn poll_revert_local_head<'smf_poll_state, 'smf_poll_context>(
        _: &'smf_poll_state mut RentToOwn<'smf_poll_state, RevertLocalHead>, 
        _: &'smf_poll_context mut RentToOwn<'smf_poll_context, Context>
    ) -> Poll<AfterRevertLocalHead, Error>;

    fn poll_add_block<'smf_poll_state, 'smf_poll_context>(
        _: &'smf_poll_state mut RentToOwn<'smf_poll_state, AddBlock>, 
        _: &'smf_poll_context mut RentToOwn<'smf_poll_context, Context>
    ) -> Poll<AfterAddBlock, Error>;
}

Required methods

fn poll_start<'smf_poll_state, 'smf_poll_context>(
_: &'smf_poll_state mut RentToOwn<'smf_poll_state, Start>,
_: &'smf_poll_context mut RentToOwn<'smf_poll_context, Context>
) -> Poll<AfterStart, Error>

Poll the future when it is in the Start state and see if it is ready to transition to a new state. If the future is ready to transition into a new state, return Ok(Async::Ready(AfterStart)). If the future is not ready to transition into a new state, return Ok(Async::NotReady). If an error is encountered, return Err(Error). The RentToOwn wrapper allows you to choose whether to take ownership of the current state or not.

fn poll_ensure_subgraph<'smf_poll_state, 'smf_poll_context>(
_: &'smf_poll_state mut RentToOwn<'smf_poll_state, EnsureSubgraph>,
_: &'smf_poll_context mut RentToOwn<'smf_poll_context, Context>
) -> Poll<AfterEnsureSubgraph, Error>

Poll the future when it is in the EnsureSubgraph state and see if it is ready to transition to a new state. If the future is ready to transition into a new state, return Ok(Async::Ready(AfterEnsureSubgraph)). If the future is not ready to transition into a new state, return Ok(Async::NotReady). If an error is encountered, return Err(Error). The RentToOwn wrapper allows you to choose whether to take ownership of the current state or not.

fn poll_load_local_head<'smf_poll_state, 'smf_poll_context>(
_: &'smf_poll_state mut RentToOwn<'smf_poll_state, LoadLocalHead>,
_: &'smf_poll_context mut RentToOwn<'smf_poll_context, Context>
) -> Poll<AfterLoadLocalHead, Error>

Poll the future when it is in the LoadLocalHead state and see if it is ready to transition to a new state. If the future is ready to transition into a new state, return Ok(Async::Ready(AfterLoadLocalHead)). If the future is not ready to transition into a new state, return Ok(Async::NotReady). If an error is encountered, return Err(Error). The RentToOwn wrapper allows you to choose whether to take ownership of the current state or not.

fn poll_poll_chain_head<'smf_poll_state, 'smf_poll_context>(
_: &'smf_poll_state mut RentToOwn<'smf_poll_state, PollChainHead>,
_: &'smf_poll_context mut RentToOwn<'smf_poll_context, Context>
) -> Poll<AfterPollChainHead, Error>

Poll the future when it is in the PollChainHead state and see if it is ready to transition to a new state. If the future is ready to transition into a new state, return Ok(Async::Ready(AfterPollChainHead)). If the future is not ready to transition into a new state, return Ok(Async::NotReady). If an error is encountered, return Err(Error). The RentToOwn wrapper allows you to choose whether to take ownership of the current state or not.

fn poll_process_blocks<'smf_poll_state, 'smf_poll_context>(
_: &'smf_poll_state mut RentToOwn<'smf_poll_state, ProcessBlocks>,
_: &'smf_poll_context mut RentToOwn<'smf_poll_context, Context>
) -> Poll<AfterProcessBlocks, Error>

Poll the future when it is in the ProcessBlocks state and see if it is ready to transition to a new state. If the future is ready to transition into a new state, return Ok(Async::Ready(AfterProcessBlocks)). If the future is not ready to transition into a new state, return Ok(Async::NotReady). If an error is encountered, return Err(Error). The RentToOwn wrapper allows you to choose whether to take ownership of the current state or not.

fn poll_vet_block<'smf_poll_state, 'smf_poll_context>(
_: &'smf_poll_state mut RentToOwn<'smf_poll_state, VetBlock>,
_: &'smf_poll_context mut RentToOwn<'smf_poll_context, Context>
) -> Poll<AfterVetBlock, Error>

Poll the future when it is in the VetBlock state and see if it is ready to transition to a new state. If the future is ready to transition into a new state, return Ok(Async::Ready(AfterVetBlock)). If the future is not ready to transition into a new state, return Ok(Async::NotReady). If an error is encountered, return Err(Error). The RentToOwn wrapper allows you to choose whether to take ownership of the current state or not.

fn poll_revert_local_head<'smf_poll_state, 'smf_poll_context>(
_: &'smf_poll_state mut RentToOwn<'smf_poll_state, RevertLocalHead>,
_: &'smf_poll_context mut RentToOwn<'smf_poll_context, Context>
) -> Poll<AfterRevertLocalHead, Error>

Poll the future when it is in the RevertLocalHead state and see if it is ready to transition to a new state. If the future is ready to transition into a new state, return Ok(Async::Ready(AfterRevertLocalHead)). If the future is not ready to transition into a new state, return Ok(Async::NotReady). If an error is encountered, return Err(Error). The RentToOwn wrapper allows you to choose whether to take ownership of the current state or not.

fn poll_add_block<'smf_poll_state, 'smf_poll_context>(
_: &'smf_poll_state mut RentToOwn<'smf_poll_state, AddBlock>,
_: &'smf_poll_context mut RentToOwn<'smf_poll_context, Context>
) -> Poll<AfterAddBlock, Error>

Poll the future when it is in the AddBlock state and see if it is ready to transition to a new state. If the future is ready to transition into a new state, return Ok(Async::Ready(AfterAddBlock)). If the future is not ready to transition into a new state, return Ok(Async::NotReady). If an error is encountered, return Err(Error). The RentToOwn wrapper allows you to choose whether to take ownership of the current state or not.

Loading content...