Type Definition graph_chain_ethereum::network_indexer::network_indexer::EnsureSubgraphFuture

type EnsureSubgraphFuture = Box<dyn Future<Item = (), Error = Error> + Send>;

Terminology used in this component:

Head / head block: The most recent block of a chain.

Local head: The block that the network indexer is at locally. We get this from the store.

Chain head: The block that the network is at. We get this from the Ethereum node(s).

Common ancestor (during a reorg): The most recent block that two versions of a chain (e.g. the locally indexed version and the latest version that the network recognizes) have in common.

When handling a reorg, this is the block after which the new version has diverged. All blocks up to and including the common ancestor remain untouched during the reorg. The blocks after the common ancestor are reverted and the blocks from the new version are added after the common ancestor.

The common ancestor is identified by traversing new blocks from a reorg back to the most recent block that we already have indexed locally. Helper types.