Trait graph_chain_ethereum::EthereumAdapterTrait

pub trait EthereumAdapterTrait: Send + Sync + 'static {
    fn url_hostname(&self) -> &str;

    fn provider(&self) -> &str;

#[must_use]    fn net_identifiers<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<EthereumNetworkIdentifier, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn latest_block(
        &self, 
        logger: &Logger
    ) -> Box<dyn Future<Item = LightEthereumBlock, Error = IngestorError> + Send + Unpin>;

    fn latest_block_header(
        &self, 
        logger: &Logger
    ) -> Box<dyn Future<Item = Block<H256>, Error = IngestorError> + Send>;

    fn load_block(
        &self, 
        logger: &Logger, 
        block_hash: H256
    ) -> Box<dyn Future<Item = LightEthereumBlock, Error = Error> + Send>;

    fn load_blocks(
        &self, 
        logger: Logger, 
        chain_store: Arc<dyn ChainStore>, 
        block_hashes: HashSet<H256>
    ) -> Box<dyn Stream<Item = LightEthereumBlock, Error = Error> + Send>;

    fn block_by_hash(
        &self, 
        logger: &Logger, 
        block_hash: H256
    ) -> Box<dyn Future<Item = Option<LightEthereumBlock>, Error = Error> + Send>;

    fn block_by_number(
        &self, 
        logger: &Logger, 
        block_number: BlockNumber
    ) -> Box<dyn Future<Item = Option<LightEthereumBlock>, Error = Error> + Send>;

    fn load_full_block(
        &self, 
        logger: &Logger, 
        block: LightEthereumBlock
    ) -> Box<dyn Future<Item = EthereumBlock, Error = IngestorError> + Send>;

    fn block_pointer_from_number(
        &self, 
        logger: &Logger, 
        block_number: BlockNumber
    ) -> Box<dyn Future<Item = BlockPtr, Error = IngestorError> + Send>;

    fn block_hash_by_block_number(
        &self, 
        logger: &Logger, 
        block_number: BlockNumber
    ) -> Box<dyn Future<Item = Option<H256>, Error = Error> + Send>;

    fn uncles(
        &self, 
        logger: &Logger, 
        block: &LightEthereumBlock
    ) -> Box<dyn Future<Item = Vec<Option<Block<H256>>>, Error = Error> + Send>;

    fn contract_call(
        &self, 
        logger: &Logger, 
        call: EthereumContractCall, 
        cache: Arc<dyn EthereumCallCache>
    ) -> Box<dyn Future<Item = Vec<Token>, Error = EthereumContractCallError> + Send>;
}

Common trait for components that watch and manage access to Ethereum.

Implementations may be implemented against an in-process Ethereum node or a remote node over RPC.

Required methods

fn url_hostname(&self) -> &str

fn provider(&self) -> &str

The provider.label from the adapter’s configuration

#[must_use]fn net_identifiers<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<EthereumNetworkIdentifier, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Ask the Ethereum node for some identifying information about the Ethereum network it is connected to.

fn latest_block(
&self,
logger: &Logger
) -> Box<dyn Future<Item = LightEthereumBlock, Error = IngestorError> + Send + Unpin>

Get the latest block, including full transactions.

fn latest_block_header(
&self,
logger: &Logger
) -> Box<dyn Future<Item = Block<H256>, Error = IngestorError> + Send>

Get the latest block, with only the header and transaction hashes.

fn load_block(
&self,
logger: &Logger,
block_hash: H256
) -> Box<dyn Future<Item = LightEthereumBlock, Error = Error> + Send>

fn load_blocks(
&self,
logger: Logger,
chain_store: Arc<dyn ChainStore>,
block_hashes: HashSet<H256>
) -> Box<dyn Stream<Item = LightEthereumBlock, Error = Error> + Send>

Load Ethereum blocks in bulk, returning results as they come back as a Stream. May use the chain_store as a cache.

fn block_by_hash(
&self,
logger: &Logger,
block_hash: H256
) -> Box<dyn Future<Item = Option<LightEthereumBlock>, Error = Error> + Send>

Find a block by its hash.

fn block_by_number(
&self,
logger: &Logger,
block_number: BlockNumber
) -> Box<dyn Future<Item = Option<LightEthereumBlock>, Error = Error> + Send>

fn load_full_block(
&self,
logger: &Logger,
block: LightEthereumBlock
) -> Box<dyn Future<Item = EthereumBlock, Error = IngestorError> + Send>

Load full information for the specified block (in particular, transaction receipts).

fn block_pointer_from_number(
&self,
logger: &Logger,
block_number: BlockNumber
) -> Box<dyn Future<Item = BlockPtr, Error = IngestorError> + Send>

Load block pointer for the specified block number.

fn block_hash_by_block_number(
&self,
logger: &Logger,
block_number: BlockNumber
) -> Box<dyn Future<Item = Option<H256>, Error = Error> + Send>

Find a block by its number, according to the Ethereum node.

Careful: don’t use this function without considering race conditions. Chain reorgs could happen at any time, and could affect the answer received. Generally, it is only safe to use this function with blocks that have received enough confirmations to guarantee no further reorgs, and where the Ethereum node is aware of those confirmations. If the Ethereum node is far behind in processing blocks, even old blocks can be subject to reorgs.

fn uncles(
&self,
logger: &Logger,
block: &LightEthereumBlock
) -> Box<dyn Future<Item = Vec<Option<Block<H256>>>, Error = Error> + Send>

Obtain all uncle blocks for a given block hash.

fn contract_call(
&self,
logger: &Logger,
call: EthereumContractCall,
cache: Arc<dyn EthereumCallCache>
) -> Box<dyn Future<Item = Vec<Token>, Error = EthereumContractCallError> + Send>

Call the function of a smart contract.

Loading content...